Combine Kaldi feature with label into one ark file
In the previous commit, the labels (utt_pdf) are read in
trainer_config.dnn.py and passed via settings()'s args parameter. The
size of the labels is restricted by protobuf's limit, which is 64 MiB.
In this commit, I combine the feats.ark and pdf.txt into one
feats_pdf.ark file by the scripts combine-ark-label.py to solve this
problem.

Also in this commit, the two shell script for running with GPU is added.

Signed-off-by: Li Peng <lipeng@unisound.com>