add op_desc.proto
In Operator design, we need a proto message to describe an Operator.
Third-party language such as python can build this protomsg and use
AddOp(const OpDesc& op_desc) of Paddle core to consturct an Op in the
Network.

The main element for an Operator is it's inputs/outputs and attributes.