Use async instead of defered for threadpool threads.
Reason:
defered means you need to call get/wait first, then
the thread starts to run.
async means the thread starts ASAP, wait/get are only
used to get the result.

The code comment says:
To wait for the completion of the task, call std::future::wait()

Hence, user would expect wait() is not a method to star the thread.
It seems more reasonable to start running when the thread is available.